Modularize couchbase writeFiles.
