+++ generators/server/files.js
@@ -29,6 +29,7 @@ const { ELASTICSEARCH } = require('../../jdl/jhipster/search-engine-types');
 const { KAFKA } = require('../../jdl/jhipster/message-broker-types');
 const { CONSUL, EUREKA } = require('../../jdl/jhipster/service-discovery-types');
 const { addSectionsCondition, mergeSections } = require('../utils');
+const { writeCouchbaseFiles } = require('./files-couchbase');
 
 /* Constants use throughout */
 const NO_DATABASE = databaseTypes.NO;
@@ -99,98 +100,6 @@ const mongoDbFiles = {
         {
           file: 'package/config/dbmigrations/InitialSetupMigration.java',
           renameTo: generator => `${generator.javaDir}config/dbmigrations/InitialSetupMigration.java`,
-        },
-      ],
-    },
-  ],
-};
-
-const couchbaseFiles = {
-  docker: [
-    {
-      path: DOCKER_DIR,
-      templates: ['couchbase-cluster.yml', 'couchbase/Couchbase.Dockerfile', 'couchbase/scripts/configure-node.sh'],
-    },
-  ],
-  serverJavaConfig: [
-    {
-      path: SERVER_MAIN_SRC_DIR,
-      templates: [
-        {
-          file: 'package/config/couchbase/CustomCouchbaseRepositoryFactory.java',
-          renameTo: generator => `${generator.javaDir}config/couchbase/CustomCouchbaseRepositoryFactory.java`,
-        },
-        {
-          file: 'package/config/couchbase/CustomCouchbaseRepositoryFactoryBean.java',
-          renameTo: generator => `${generator.javaDir}config/couchbase/CustomCouchbaseRepositoryFactoryBean.java`,
-        },
-        {
-          file: 'package/config/couchbase/CustomCouchbaseRepositoryQuery.java',
-          renameTo: generator => `${generator.javaDir}config/couchbase/CustomCouchbaseRepositoryQuery.java`,
-        },
-        {
-          file: 'package/config/couchbase/CustomN1qlQueryCreator.java',
-          renameTo: generator => `${generator.javaDir}config/couchbase/CustomN1qlQueryCreator.java`,
-        },
-        {
-          file: 'package/config/couchbase/CustomN1qlRepositoryQueryExecutor.java',
-          renameTo: generator => `${generator.javaDir}config/couchbase/CustomN1qlRepositoryQueryExecutor.java`,
-        },
-        {
-          file: 'package/config/couchbase/package-info.java',
-          renameTo: generator => `${generator.javaDir}config/couchbase/package-info.java`,
-        },
-        {
-          file: 'package/repository/JHipsterCouchbaseRepository.java',
-          renameTo: generator => `${generator.javaDir}repository/JHipsterCouchbaseRepository.java`,
-        },
-      ],
-    },
-    {
-      condition: generator => !shouldSkipUserManagement(generator) && generator.authenticationType === SESSION && !generator.reactive,
-      path: SERVER_MAIN_SRC_DIR,
-      templates: [
-        {
-          file: 'package/repository/PersistentTokenRepository_couchbase.java',
-          renameTo: generator => `${generator.javaDir}repository/PersistentTokenRepository.java`,
-        },
-      ],
-    },
-    {
-      condition: generator => generator.searchEngine === COUCHBASE,
-      path: SERVER_TEST_SRC_DIR,
-      templates: [
-        {
-          file: 'package/repository/JHipsterCouchbaseRepositoryTest.java',
-          renameTo: generator => `${generator.testDir}repository/JHipsterCouchbaseRepositoryTest.java`,
-        },
-      ],
-    },
-  ],
-  serverResource: [
-    {
-      condition: generator => generator.databaseTypeCouchbase,
-      path: SERVER_MAIN_RES_DIR,
-      templates: ['config/couchmove/changelog/V0__create_indexes.n1ql'],
-    },
-    {
-      condition: generator => !generator.skipUserManagement || generator.authenticationType === OAUTH2,
-      path: SERVER_MAIN_RES_DIR,
-      templates: [
-        'config/couchmove/changelog/V0.1__initial_setup/ROLE_ADMIN.json',
-        'config/couchmove/changelog/V0.1__initial_setup/ROLE_USER.json',
-        'config/couchmove/changelog/V0.1__initial_setup/user__admin.json',
-        'config/couchmove/changelog/V0.1__initial_setup/user__user.json',
-      ],
-    },
-  ],
-  serverTestFw: [
-    {
-      path: SERVER_TEST_SRC_DIR,
-      templates: [
-        {
-          file: 'package/CouchbaseTestContainerExtension.java',
-          renameTo: generator => `${generator.testDir}CouchbaseTestContainerExtension.java`,
         },
       ],
     },
@@ -1801,7 +1710,6 @@ const serverFiles = mergeSections(
   addSectionsCondition(h2Files, context => context.devDatabaseTypeH2Any),
   addSectionsCondition(liquibaseFiles, context => context.databaseTypeSql),
   addSectionsCondition(mongoDbFiles, context => context.databaseTypeMongodb),
-  addSectionsCondition(couchbaseFiles, context => context.databaseTypeCouchbase),
   addSectionsCondition(neo4jFiles, context => context.databaseTypeNeo4j),
   addSectionsCondition(cassandraFiles, context => context.databaseTypeCassandra)
 );
@@ -1828,6 +1736,8 @@ function writeFiles() {
     writeFiles() {
       return this.writeFilesToDisk(serverFiles);
     },
+
+    ...writeCouchbaseFiles(),
   };
 }
